CLI処理用のMemoCLIクラスを作成、コマンドライン上のメモ入力をMemoCLIクラスの責務とした
